move the list to another module and improve the cons and cons_end
