+++ numbat/src/list.rs
@@ -1,0 +1,155 @@
+use std::{collections::VecDeque, fmt, sync::Arc};
+
+use crate::{value::Value, RuntimeError};
+
+/// Reference counted list / list view
+#[derive(Clone, Eq)]
+pub struct NumbatList<T> {
+    /// The original alloc shared between all values
+    alloc: Arc<VecDeque<T>>,
+    /// The indexes accessible to us. If `None` we own the whole allocation
+    view: Option<(usize, usize)>,
+}
+
+impl<T: fmt::Debug + Clone> fmt::Debug for NumbatList<T> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_list().entries(self.iter()).finish()
+    }
+}
+
+impl<T: PartialEq> PartialEq for NumbatList<T> {
+    fn eq(&self, other: &Self) -> bool {
+        // Best case scenario, the slice don't have the same len and we can early exit
+        if self.len() != other.len() {
+            return false;
+        }
+        // Second best case scenario, the slice comes from the same allocation
+        // and have the same view => they are equal
+        if Arc::as_ptr(&self.alloc) == Arc::as_ptr(&other.alloc) && self.view == other.view {
+            true
+        } else {
+            // Worst case scenario, we need to compare all the elements one by one
+            self.iter().zip(other.iter()).all(|(l, r)| l == r)
+        }
+    }
+}
+
+impl<T> NumbatList<T> {
+    pub fn len(&self) -> usize {
+        if let Some(view) = self.view {
+            view.1 - view.0
+        } else {
+            self.alloc.len()
+        }
+    }
+
+    pub fn with_capacity(capacity: usize) -> Self {
+        Self {
+            alloc: Arc::new(VecDeque::with_capacity(capacity)),
+            view: None,
+        }
+    }
+
+    pub fn front(&self) -> Option<&T> {
+        if let Some(view) = self.view {
+            self.alloc.get(view.0)
+        } else {
+            self.alloc.front()
+        }
+    }
+
+    /// Advance the view you have of the list by one.
+    pub fn advance_view(&mut self) -> Result<(), RuntimeError> {
+        if self.len() == 0 {
+            return Err(RuntimeError::EmptyList);
+        }
+        if let Some(view) = &mut self.view {
+            view.0 += 1;
+            // should be ensured by the if above
+            debug_assert!(view.0 <= view.1);
+        } else {
+            self.view = Some((1, self.len()));
+        }
+        Ok(())
+    }
+
+    pub fn iter(&self) -> impl Iterator<Item = &T> {
+        self.alloc
+            .iter()
+            .skip(self.view.map_or(0, |(start, _end)| start))
+            .take(
+                self.view
+                    .map_or(self.alloc.len(), |(start, end)| end - start),
+            )
+    }
+}
+
+impl<T: Clone> NumbatList<T> {
+    fn make_mut(&mut self) -> &mut VecDeque<T> {
+        if Arc::strong_count(&self.alloc) != 1 {
+            // If someone else is using the list we must clone it
+            self.alloc = Arc::new(self.iter().cloned().collect());
+            self.view = None;
+        }
+        // With our usage, this should never allocate since we know we're the only
+        // one holding a reference to this `Arc` and we don't use weak references.
+        Arc::make_mut(&mut self.alloc)
+    }
+
+    /// Allocate iif the list being used by another value at the same time
+    pub fn push_front(&mut self, element: T) {
+        let mut view = self.view;
+        let inner = self.make_mut();
+        if let Some((start, end)) = &mut view {
+            // if we were alone on the allocation and had a view of the inner allocation
+            // we can keep the allocation and overwrite the start-1 element.
+            // but we need to take care of the special case where the start is 0.
+            if *start == 0 {
+                inner.push_front(element);
+                *end += 1;
+            } else {
+                *start -= 1;
+                inner[*start] = element;
+            }
+        } else {
+            inner.push_front(element);
+        }
+        self.view = view;
+    }
+
+    /// Allocate iif the list being used by another value at the same time
+    pub fn push_back(&mut self, element: T) {
+        let mut view = self.view;
+        let inner = self.make_mut();
+        if let Some((_start, end)) = &mut view {
+            // if we were alone on the allocation and had a view of the inner allocation
+            // we can keep the allocation and overwrite the end+1 element.
+            // but we need to take care of the special case where the end is `alloc.len()`.
+            if *end == inner.len() {
+                inner.push_back(element);
+                *end += 1;
+            } else {
+                *end += 1;
+                inner[*end] = element;
+            }
+        } else {
+            inner.push_back(element);
+        }
+        self.view = view;
+    }
+}
+
+impl From<NumbatList<Value>> for Value {
+    fn from(list: NumbatList<Value>) -> Self {
+        Value::List(list)
+    }
+}
+
+impl From<VecDeque<Value>> for Value {
+    fn from(list: VecDeque<Value>) -> Self {
+        Value::List(NumbatList {
+            alloc: Arc::new(list),
+            view: None,
+        })
+    }
+}
